Refactor: 게시글 업데이트 서비스 트랜잭션 처리
1. 게시글 작업은 트랜잭션 내에서 처리되도록 함
2. 게시글 작업이 성공했을 때에만 이미지 삭제되도록 트랜잭션 매니저 사용
2-1. 만약 게시글 작업이 실패하면 이미지 삭제 작업은 진행되지 않음
2-2. 게시글 작업 트랜잭션이 커밋된 후에 이미지 삭제 작업이 진행되도록 하여 게시글은 롤백됐는데 이미지는 삭제된 상황을 방지
