+++ src/main/java/com/jandi/plan_backend/util/CommunityUtil.java
@@ -10,8 +10,10 @@ import com.jandi.plan_backend.image.repository.ImageRepository;
 import com.jandi.plan_backend.image.service.ImageService;
 import com.jandi.plan_backend.user.entity.User;
 import com.jandi.plan_backend.user.repository.UserRepository;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Propagation;
 
 import java.util.HashSet;
 import java.util.List;
@@ -23,7 +25,6 @@ import java.util.regex.Pattern;
 public class CommunityUtil {
     private final UserRepository userRepository;
     private final ImageRepository imageRepository;
-    private final ImageService imageService;
     private final CommunityLikeRepository communityLikeRepository;
 
     @Value("${image-prefix}")
@@ -32,11 +33,9 @@ public class CommunityUtil {
     public CommunityUtil(
             UserRepository userRepository,
             ImageRepository imageRepository,
-            ImageService imageService,
             CommunityLikeRepository communityLikeRepository) {
         this.userRepository = userRepository;
         this.imageRepository = imageRepository;
-        this.imageService = imageService;
         this.communityLikeRepository = communityLikeRepository;
     }
 
@@ -79,32 +78,5 @@ public class CommunityUtil {
             throw new RuntimeException(e);
         }
         return preview.toString();
-    }
-
-    /**
-     * 게시글 내용에서 실제 사용 중인 이미지 파일명을 추출합니다.
-     * 예: "https://storage.googleapis.com/plan-storage/encodedFileName.jpg"에서 "encodedFileName.jpg" 추출
-     */
-    public Set<String> extractImageFileNamesFromContent(String content) {
-        Set<String> fileNames = new HashSet<>();
-        Pattern pattern = Pattern.compile(prefix + "([^\"\\s]+)");
-        Matcher matcher = pattern.matcher(content);
-        while (matcher.find()) {
-            fileNames.add(matcher.group(1));
-        }
-        return fileNames;
-    }
-
-    /**
-     * 게시글과 연결된 이미지 중, 게시글 내용에 포함되지 않은 이미지를 삭제합니다.
-     */
-    public void cleanupUnusedImages(Community post) {
-        Set<String> usedFileNames = extractImageFileNamesFromContent(post.getContents());
-        List<Image> images = imageRepository.findAllByTargetTypeAndTargetId("community", post.getPostId());
-        for (Image image : images) {
-            if (!usedFileNames.contains(image.getImageUrl())) {
-                imageService.deleteImage(image.getImageId());
-            }
-        }
     }
 }
